feat: add SVG layer
Reference coordinates are CSS px.
SVG only uses CSS px, while canvas scales depending on devicePixelRatio.

It means that the number of pixels in the canvas will be higher on HD or
retina screens (eg. every CSS pixel is rendered by 4 or 9 physical pixels)
When a browser zoom is applied, the number of pixels in the canvas will
also be adapted. But as the image takes all the screen (100vh for
<v-content>), the apparent resolution stays unchanged when zooming or
unzooming.